fix memory leak in -E mode
When running with -fsanitize=leak enabled nasm prints this error:

Direct leak of 124 byte(s) in 20 object(s) allocated from:
    #0 0x7e96126fd9c7 in malloc ../../../../src/libsanitizer/asan/asan_malloc_linux.cpp:69
    #1 0x55a08fa0f6b5 in nasm_malloc nasmlib/alloc.c:25
    #2 0x55a08f90db3a in detoken asm/preproc.c:2219
    #3 0x55a08f949b38 in pp_getline asm/preproc.c:8815
    #4 0x55a08f8ab33e in main asm/nasm.c:602
    #5 0x7e9611a2a1c9 in __libc_start_call_main ../sysdeps/nptl/libc_start_call_main.h:58
    #6 0x7e9611a2a28a in __libc_start_main_impl ../csu/libc-start.c:360
    #7 0x55a08f8a6e84 in _start	(/home/ivan/d/nasm/nasm+0x34be84)

This is reproducible on tests that do preprocessing for example
weirdpaste.asm test.

The problem is caused by the fact that the line returned by pp_getline
isn't freed in main function.

Signed-off-by: Ivan Sorokin <vanyacpp@gmail.com>