Fix PC-relative expression relocations
This change fixes relocations which are emitted as a result of PC-relative assembler expressions (of the form `XYZ - $`). Previously, these expressions were evaluated down to offsets, and the offset itself (which could be negative!) was passed onto the output backend, instead of the actual absolute address. In particular, this could cause e.g. the Mach-O backend to emit a relocation against garbage symbols, as attempts to resolve the symbol based on the invalid address would not find the actual symbol being referenced in the original code. This was fixed in this patch by making the address absolute again before passing it onto the backend, as well as setting `data->relbase` so that the backend can properly compute the relative offset.