preproc: Fix makefile dependencies for included files.
A bug in 169ac7c152ee13ed0c470ceb3371e9afb10e9a60 caused inc_fopen to
not add included files to the dependency list when using the -MD file
option.  The reason is that the included files would only be added to
the list on a hash table miss, but since the dependency list is generated
in the final pass when the hash table is already fully populated, no
include files would actually be added.

The fix is to always add an included file to the dependency list,
regardless of the hash table hit status.

Signed-off-by: knut st. osmundsen <bird-nasm@anduin.net>